# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "认知知识库系统 API",
    "description": "文档上传 → 知识抽取 → 图谱可视化 → 知识检索（PRD V1.3 P0 核心闭环）",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/graph/overview": {
      "get": {
        "tags": [
          "知识图谱"
        ],
        "summary": "图谱总览（点线数据）",
        "operationId": "graph_overview_api_v1_graph_overview_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "description": "返回的边数上限（采样保护）",
              "default": 500,
              "title": "Limit"
            },
            "description": "返回的边数上限（采样保护）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphOverviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/node/{entity_id}": {
      "get": {
        "tags": [
          "知识图谱"
        ],
        "summary": "节点详情（关系+邻居+证据句）",
        "operationId": "node_detail_api_v1_graph_node__entity_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/fulltext": {
      "get": {
        "tags": [
          "知识检索"
        ],
        "summary": "全文检索",
        "operationId": "search_fulltext_api_v1_search_fulltext_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Q"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FulltextHitOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/entities": {
      "get": {
        "tags": [
          "知识检索"
        ],
        "summary": "实体检索",
        "operationId": "search_entities_api_v1_search_entities_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Q"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_EntityHitOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ontology/entity-types": {
      "get": {
        "tags": [
          "本体管理"
        ],
        "summary": "实体类型列表（图例数据）",
        "operationId": "list_entity_types_api_v1_ontology_entity_types_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityTypesOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ontology/relation-types": {
      "get": {
        "tags": [
          "本体管理"
        ],
        "summary": "关系类型列表",
        "operationId": "list_relation_types_api_v1_ontology_relation_types_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationTypesOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "color": {
            "type": "string",
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "color"
        ],
        "title": "CategoryOut",
        "description": "实体类型图例（节点着色依据，GRA-03）。"
      },
...
      "EntityHitOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "additionalProperties": true,
            "type": "object",
            "title": "Type"
          },
          "frequency": {
            "type": "integer",
            "title": "Frequency"
          },
          "relation_count": {
            "type": "integer",
            "title": "Relation Count"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "frequency",
          "relation_count",
          "document_count"
        ],
        "title": "EntityHitOut"
      },
      "EntityTypeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "color",
          "description",
          "created_at"
        ],
        "title": "EntityTypeOut"
      },
      "EntityTypesOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntityTypeOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "EntityTypesOut"
      },
      "EvidenceOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "sentence": {
            "type": "string",
            "title": "Sentence"
          },
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "document_filename": {
            "type": "string",
            "title": "Document Filename"
          },
          "segment_id": {
            "type": "integer",
            "title": "Segment Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sentence",
          "document_id",
          "document_filename",
          "segment_id"
        ],
        "title": "EvidenceOut"
      },
      "FulltextHitOut": {
        "properties": {
          "document_id": {
            "type": "integer",
            "title": "Document Id"
          },
          "document_filename": {
            "type": "string",
            "title": "Document Filename"
          },
          "segment_id": {
            "type": "integer",
            "title": "Segment Id"
          },
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          },
          "matched": {
            "type": "boolean",
            "title": "Matched"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "document_filename",
          "segment_id",
          "seq",
          "snippet",
          "matched"
        ],
        "title": "FulltextHitOut"
      },
      "GraphEdgeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source": {
            "type": "integer",
            "title": "Source"
          },
          "target": {
            "type": "integer",
            "title": "Target"
          },
          "relation_type_name": {
            "type": "string",
            "title": "Relation Type Name"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "target",
          "relation_type_name",
          "weight"
        ],
        "title": "GraphEdgeOut"
      },
      "GraphNodeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "value": {
            "type": "integer",
            "title": "Value"
          },
          "category": {
            "type": "integer",
            "title": "Category"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "type_name": {
            "type": "string",
            "title": "Type Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "value",
          "category",
          "color",
          "type_name"
        ],
        "title": "GraphNodeOut"
      },
      "GraphOverviewOut": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNodeOut"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdgeOut"
            },
            "type": "array",
            "title": "Edges"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryOut"
            },
            "type": "array",
            "title": "Categories"
          },
          "relation_types": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Relation Types"
          },
          "total_nodes": {
            "type": "integer",
            "title": "Total Nodes"
          },
          "total_edges": {
            "type": "integer",
            "title": "Total Edges"
          },
          "sampled": {
            "type": "boolean",
            "title": "Sampled"
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "edges",
          "categories",
          "relation_types",
          "total_nodes",
          "total_edges",
          "sampled"
        ],
        "title": "GraphOverviewOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NeighborOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type_name": {
            "type": "string",
            "title": "Type Name"
          },
          "color": {
            "type": "string",
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type_name",
          "color"
        ],
        "title": "NeighborOut"
      },
      "NodeDetailOut": {
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/NodeEntityOut"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/NodeRelationOut"
            },
            "type": "array",
            "title": "Relations"
          },
          "evidences": {
            "items": {
              "$ref": "#/components/schemas/EvidenceOut"
            },
            "type": "array",
            "title": "Evidences"
          }
        },
        "type": "object",
        "required": [
          "entity",
          "relations",
          "evidences"
        ],
        "title": "NodeDetailOut"
      },
      "NodeEntityOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "additionalProperties": true,
            "type": "object",
            "title": "Type"
          },
          "frequency": {
            "type": "integer",
            "title": "Frequency"
          },
          "attributes": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attributes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "frequency",
          "attributes"
        ],
        "title": "NodeEntityOut"
      },
      "NodeRelationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "relation_type_name": {
            "type": "string",
            "title": "Relation Type Name"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          },
          "other": {
            "$ref": "#/components/schemas/NeighborOut"
          }
        },
        "type": "object",
        "required": [
          "id",
          "relation_type_name",
          "direction",
          "weight",
          "other"
        ],
        "title": "NodeRelationOut"
      },
...
      "Page_EntityHitOut_": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntityHitOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "page_size",
          "items"
        ],
        "title": "Page[EntityHitOut]"
      },
      "Page_FulltextHitOut_": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FulltextHitOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "page_size",
          "items"
        ],
        "title": "Page[FulltextHitOut]"
      },
...
      "RelationTypeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_at"
        ],
        "title": "RelationTypeOut"
      },
      "RelationTypesOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RelationTypeOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "RelationTypesOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}